--- src/components/MarkdownOutput/helpers/isFunction.js
@@ -1,3 +1,0 @@
-export default function isFunction(value) {
-    return typeof value === 'function';
-}
