--- src/components/MarkdownOutput/helpers/removeListItemParagraphWrap.js
@@ -1,5 +1,0 @@
-export default function removeListItemParagraphWrap(node) {
-    return node.children.reduce((newChildrenList, child) => {
-        return newChildrenList.concat(child.type === 'paragraph' ? child.children || [] : [child]);
-    }, []);
-}
