--- src/components/MarkdownOutput/hooks/index.js
@@ -1,2 +1,0 @@
-export { default as useMarkdownParser } from './useMarkdownParser';
-export { default as useSyntaxTreeJsxTranform } from './useSyntaxTreeJsxTranform';
