+++ src/components/MarkdownOutput/index.d.ts
@@ -4,9 +4,6 @@ import { BaseProps } from '../types';
 export interface MarkdownOutputProps extends BaseProps {
     id?: string;
     renderers?: object;
-    linkTarget?: string;
-    transformLinkUri?: () => void;
-    transformImageUri?: () => void;
     plugins?: object[];
     parserOptions?: object;
     source: string;
