+++ src/components/MarkdownOutput/index.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { useMarkdownParser, useSyntaxTreeJsxTranform } from './hooks';
+import useMarkdownToReact from './hooks/useMarkdownToReact';
+import defaultRenderers from './renderers';
 
 export default function MarkdownOutput(props) {
     const {
@@ -11,23 +12,13 @@ export default function MarkdownOutput(props) {
         renderers: customRenderers,
         parserOptions,
         plugins,
-        linkTarget,
-        transformLinkUri,
-        transformImageUri,
     } = props;
 
     const source = sourceInProps || '';
-    const rawAbstractSyntaxTree = useMarkdownParser(source, parserOptions, plugins);
-    const result = useSyntaxTreeJsxTranform(
-        rawAbstractSyntaxTree,
-        {
-            linkTarget,
-            transformLinkUri,
-            transformImageUri,
-        },
-        customRenderers,
-    );
-
+    const result = useMarkdownToReact(source, parserOptions, plugins, {
+        ...defaultRenderers,
+        ...customRenderers,
+    });
     return (
         <div id={id} className={className} style={style}>
             {result}
@@ -50,23 +41,6 @@ MarkdownOutput.propTypes = {
      *  varies based on the type of node.
      */
     renderers: PropTypes.object,
-    /** Sets the default target attribute for links. Default is undefined (no target attribute). */
-    linkTarget: PropTypes.string,
-    /**
-     *  Function that gets called for each encountered link. The returned value is
-     *  used in place of the original URI. If you want to disable the default transformer, pass
-     *  null to this option.
-     *  @param {string} uri
-     *  @returns {string}
-     */
-    transformLinkUri: PropTypes.func,
-    /**
-     *  Function that gets called for each encountered image.
-     *  The returned value is used in place of the original image src.
-     *  @param {string} uri
-     *  @returns {string}
-     */
-    transformImageUri: PropTypes.func,
     /**
      *  An array of unified/remark parser plugins. If you need to pass options
      *  to the plugin, pass an array with two elements, the first being the
@@ -84,9 +58,6 @@ MarkdownOutput.defaultProps = {
     className: undefined,
     style: undefined,
     renderers: undefined,
-    linkTarget: undefined,
-    transformLinkUri: undefined,
-    transformImageUri: undefined,
     plugins: [],
     parserOptions: undefined,
 };
